Add early version of synchronization, fix sending
Add locks in parsingMessages, this is the right way, no deadlocks or
hangs.
Fix the sending, now everything should be sent correctly and I should
receive all the responses. I still have to check if ack is ok and add
the watcher.
I need to deliver (write to buffer) only when all the conditions are met
and not before and I should start using the count on the expected to
check if a process has failed.